+++ tests/rest/datacube/test_datacube.py
@@ -4,7 +4,10 @@ General cube method tests against both
 - 1.0.0-style DataCube
 
 """
+
+import contextlib
 import pathlib
+import re
 from datetime import date, datetime
 from unittest import mock
 
@@ -14,6 +17,7 @@ import requests
 import shapely
 import shapely.geometry
 
+from openeo.api.process import Parameter
 from openeo.rest import BandMathException, OpenEoClientException
 from openeo.rest._testing import build_capabilities
 from openeo.rest.connection import Connection
@@ -98,26 +102,39 @@ def test_filter_temporal_basic_extent(s2cube):
     assert graph['arguments']['extent'] == ["2016-01-01", "2016-03-10"]
 
 
-@pytest.mark.parametrize("args,kwargs,extent", [
-    ((), {}, [None, None]),
-    (("2016-01-01",), {}, ["2016-01-01", None]),
-    (("2016-01-01", "2016-03-10"), {}, ["2016-01-01", "2016-03-10"]),
-    ((date(2016, 1, 1), date(2016, 3, 10)), {}, ["2016-01-01", "2016-03-10"]),
-    ((datetime(2016, 1, 1, 12, 34), datetime(2016, 3, 10, 23, 45)), {},
-     ["2016-01-01T12:34:00Z", "2016-03-10T23:45:00Z"]),
-    ((), {"start_date": "2016-01-01", "end_date": "2016-03-10"}, ["2016-01-01", "2016-03-10"]),
-    ((), {"start_date": "2016-01-01"}, ["2016-01-01", None]),
-    ((), {"end_date": "2016-03-10"}, [None, "2016-03-10"]),
-    ((), {"start_date": date(2016, 1, 1), "end_date": date(2016, 3, 10)}, ["2016-01-01", "2016-03-10"]),
-    ((), {"start_date": datetime(2016, 1, 1, 12, 34), "end_date": datetime(2016, 3, 10, 23, 45)},
-     ["2016-01-01T12:34:00Z", "2016-03-10T23:45:00Z"]),
-    ((), {"extent": ("2016-01-01", "2016-03-10")}, ["2016-01-01", "2016-03-10"]),
-    ((), {"extent": ("2016-01-01", None)}, ["2016-01-01", None]),
-    ((), {"extent": (None, "2016-03-10")}, [None, "2016-03-10"]),
-    ((), {"extent": (date(2016, 1, 1), date(2016, 3, 10))}, ["2016-01-01", "2016-03-10"]),
-    ((), {"extent": (datetime(2016, 1, 1, 12, 34), datetime(2016, 3, 10, 23, 45))},
-     ["2016-01-01T12:34:00Z", "2016-03-10T23:45:00Z"]),
-])
+@pytest.mark.parametrize(
+    "args,kwargs,extent",
+    [
+        ((), {}, [None, None]),
+        (("2016-01-01", "2016-03-10"), {}, ["2016-01-01", "2016-03-10"]),
+        ((("2016-01-01", "2016-03-10"),), {}, ["2016-01-01", "2016-03-10"]),
+        ((["2016-01-01", "2016-03-10"],), {}, ["2016-01-01", "2016-03-10"]),
+        ((date(2016, 1, 1), date(2016, 3, 10)), {}, ["2016-01-01", "2016-03-10"]),
+        (
+            (datetime(2016, 1, 1, 12, 34), datetime(2016, 3, 10, 23, 45)),
+            {},
+            ["2016-01-01T12:34:00Z", "2016-03-10T23:45:00Z"],
+        ),
+        ((), {"start_date": "2016-01-01", "end_date": "2016-03-10"}, ["2016-01-01", "2016-03-10"]),
+        ((), {"start_date": "2016-01-01"}, ["2016-01-01", None]),
+        ((), {"end_date": "2016-03-10"}, [None, "2016-03-10"]),
+        ((), {"start_date": date(2016, 1, 1), "end_date": date(2016, 3, 10)}, ["2016-01-01", "2016-03-10"]),
+        (
+            (),
+            {"start_date": datetime(2016, 1, 1, 12, 34), "end_date": datetime(2016, 3, 10, 23, 45)},
+            ["2016-01-01T12:34:00Z", "2016-03-10T23:45:00Z"],
+        ),
+        ((), {"extent": ("2016-01-01", "2016-03-10")}, ["2016-01-01", "2016-03-10"]),
+        ((), {"extent": ("2016-01-01", None)}, ["2016-01-01", None]),
+        ((), {"extent": (None, "2016-03-10")}, [None, "2016-03-10"]),
+        ((), {"extent": (date(2016, 1, 1), date(2016, 3, 10))}, ["2016-01-01", "2016-03-10"]),
+        (
+            (),
+            {"extent": (datetime(2016, 1, 1, 12, 34), datetime(2016, 3, 10, 23, 45))},
+            ["2016-01-01T12:34:00Z", "2016-03-10T23:45:00Z"],
+        ),
+    ],
+)
 def test_filter_temporal_generic(s2cube, args, kwargs, extent):
     im = s2cube.filter_temporal(*args, **kwargs)
     graph = _get_leaf_node(im)
@@ -389,6 +406,28 @@ def test_filter_temporal_extent_single_date_string(s2cube: DataCube, api_version
     cube_extent: DataCube = s2cube.filter_temporal(extent=extent)
     flat_graph_extent = cube_extent.flat_graph()
     assert flat_graph_extent["filtertemporal1"]["arguments"]["extent"] == expected
+
+
+@pytest.mark.parametrize(
+    ["arg", "expect_failure"],
+    [
+        ("2024-09-24", True),
+        ("2024-09", True),
+        ("2024", True),
+        (("2024-09-01", "2024-09-10"), False),
+        (["2024-09-01", "2024-09-10"], False),
+        (Parameter.temporal_interval("window"), False),
+    ],
+)
+def test_filter_temporal_single_arg(s2cube: DataCube, arg, expect_failure):
+    if expect_failure:
+        context = pytest.raises(
+            OpenEoClientException, match="filter_temporal.*with a single string argument.*is ambiguous.*"
+        )
+    else:
+        context = contextlib.nullcontext()
+    with context:
+        _ = s2cube.filter_temporal(arg)
 
 
 def test_max_time(s2cube, api_version):
@@ -672,13 +711,14 @@ class TestExecuteBatch:
     @pytest.mark.parametrize(
         ["save_result_format", "execute_format", "expected"],
         [
-            ("GTiff", "GTiff", "GTiff"),
+            (None, None, "GTiff"),
+            (None, "GTiff", "GTiff"),
             ("GTiff", None, "GTiff"),
-            ("NetCDF", "NetCDF", "NetCDF"),
+            (None, "NetCDF", "NetCDF"),
             ("NetCDF", None, "NetCDF"),
         ],
     )
-    def test_create_job_existing_save_result(
+    def test_save_result_and_create_job_at_most_one_with_format(
         self,
         s2cube,
         get_create_job_pg,
@@ -686,7 +726,10 @@ class TestExecuteBatch:
         execute_format,
         expected,
     ):
-        cube = s2cube.save_result(format=save_result_format)
+        cube = s2cube
+        if save_result_format:
+            cube = cube.save_result(format=save_result_format)
+
         cube.create_job(out_format=execute_format)
         pg = get_create_job_pg()
         assert set(pg.keys()) == {"loadcollection1", "saveresult1"}
@@ -702,14 +745,33 @@ class TestExecuteBatch:
 
     @pytest.mark.parametrize(
         ["save_result_format", "execute_format"],
-        [("NetCDF", "GTiff"), ("GTiff", "NetCDF")],
+        [
+            ("NetCDF", "NetCDF"),
+            ("GTiff", "NetCDF"),
+        ],
     )
-    def test_create_job_existing_save_result_incompatible(
-        self, s2cube, save_result_format, execute_format
-    ):
+    def test_save_result_and_create_job_both_with_format(self, s2cube, save_result_format, execute_format):
         cube = s2cube.save_result(format=save_result_format)
-        with pytest.raises(ValueError):
+        with pytest.raises(
+            OpenEoClientException,
+            match=re.escape(
+                "DataCube.create_job() with explicit output format 'NetCDF',"
+                " but the process graph already has `save_result` node(s)"
+                " which is ambiguous and should not be combined."
+            ),
+        ):
             cube.create_job(out_format=execute_format)
+
+    @pytest.mark.parametrize(
+        ["auto_add_save_result", "process_ids"],
+        [
+            (True, {"load_collection", "save_result"}),
+            (False, {"load_collection"}),
+        ],
+    )
+    def test_create_job_auto_add_save_result(self, s2cube, dummy_backend, auto_add_save_result, process_ids):
+        s2cube.create_job(auto_add_save_result=auto_add_save_result)
+        assert set(n["process_id"] for n in dummy_backend.get_pg().values()) == process_ids
 
     def test_execute_batch_defaults(self, s2cube, get_create_job_pg, recwarn, caplog):
         s2cube.execute_batch()
@@ -770,13 +832,14 @@ class TestExecuteBatch:
     @pytest.mark.parametrize(
         ["save_result_format", "execute_format", "expected"],
         [
-            ("GTiff", "GTiff", "GTiff"),
+            (None, None, "GTiff"),
+            (None, "GTiff", "GTiff"),
             ("GTiff", None, "GTiff"),
-            ("NetCDF", "NetCDF", "NetCDF"),
+            (None, "NetCDF", "NetCDF"),
             ("NetCDF", None, "NetCDF"),
         ],
     )
-    def test_execute_batch_existing_save_result(
+    def test_save_result_and_execute_batch_at_most_one_with_format(
         self,
         s2cube,
         get_create_job_pg,
@@ -784,7 +847,9 @@ class TestExecuteBatch:
         execute_format,
         expected,
     ):
-        cube = s2cube.save_result(format=save_result_format)
+        cube = s2cube
+        if save_result_format:
+            cube = cube.save_result(format=save_result_format)
         cube.execute_batch(out_format=execute_format)
         pg = get_create_job_pg()
         assert set(pg.keys()) == {"loadcollection1", "saveresult1"}
@@ -800,14 +865,61 @@ class TestExecuteBatch:
 
     @pytest.mark.parametrize(
         ["save_result_format", "execute_format"],
-        [("NetCDF", "GTiff"), ("GTiff", "NetCDF")],
+        [
+            ("NetCDF", "NetCDF"),
+            ("GTiff", "NetCDF"),
+        ],
     )
     def test_execute_batch_existing_save_result_incompatible(
         self, s2cube, save_result_format, execute_format
     ):
         cube = s2cube.save_result(format=save_result_format)
-        with pytest.raises(ValueError):
+        with pytest.raises(
+            OpenEoClientException,
+            match=re.escape(
+                "DataCube.execute_batch() with explicit output format 'NetCDF',"
+                " but the process graph already has `save_result` node(s)"
+                " which is ambiguous and should not be combined."
+            ),
+        ):
             cube.execute_batch(out_format=execute_format)
+
+    @pytest.mark.parametrize(
+        ["save_result_format", "execute_output_file", "expected"],
+        [
+            (None, None, "GTiff"),
+            (None, "result.tiff", "GTiff"),
+            ("GTiff", None, "GTiff"),
+            ("GTiff", "result.csv", "GTiff"),
+            (None, "result.nc", "netCDF"),
+            ("NetCDF", None, "NetCDF"),
+            ("NetCDF", "result.csv", "NetCDF"),
+            (None, "result.csv", "CSV"),
+        ],
+    )
+    def test_save_result_and_execute_batch_weak_format(
+        self,
+        s2cube,
+        get_create_job_pg,
+        save_result_format,
+        execute_output_file,
+        expected,
+    ):
+        cube = s2cube
+        if save_result_format:
+            cube = cube.save_result(format=save_result_format)
+        cube.execute_batch(outputfile=execute_output_file)
+        pg = get_create_job_pg()
+        assert set(pg.keys()) == {"loadcollection1", "saveresult1"}
+        assert pg["saveresult1"] == {
+            "process_id": "save_result",
+            "arguments": {
+                "data": {"from_node": "loadcollection1"},
+                "format": expected,
+                "options": {},
+            },
+            "result": True,
+        }
 
     def test_save_result_format_options_vs_create_job(elf, s2cube, get_create_job_pg):
         """https://github.com/Open-EO/openeo-python-client/issues/433"""
@@ -840,6 +952,17 @@ class TestExecuteBatch:
             },
             "result": True,
         }
+
+    @pytest.mark.parametrize(
+        ["auto_add_save_result", "process_ids"],
+        [
+            (True, {"load_collection", "save_result"}),
+            (False, {"load_collection"}),
+        ],
+    )
+    def test_execute_batch_auto_add_save_result(self, s2cube, dummy_backend, auto_add_save_result, process_ids):
+        s2cube.execute_batch(auto_add_save_result=auto_add_save_result)
+        assert set(n["process_id"] for n in dummy_backend.get_pg().values()) == process_ids
 
 
 class TestDataCubeValidation: